Add trailing spaces to each command line that are missing them.
The first is necessary for the command to run correctly, and while the second isn't strictly necessary, it does mean that if we add a new line, its less likely to be broken by careless forgetting of the space. Kind of like adding the last comma in an array definition to make future work easier.